refactor(test): P5-2 follow-up — extract CG gradient/pattern cluster
Continue the test_canvas.cpp split started in P5-2's first cut. Extracts
the 628-line CoreGraphics gradient + pattern cluster (pulp #1524) into a
sibling TU test/test_canvas_cg_gradients.cpp.

* test_canvas.cpp: 3,363 → 2,735 lines
* New target pulp-test-canvas-cg-gradients: 654 lines, 9 TEST_CASEs
* Apple-only TU — every test path drives CoreGraphicsCanvas directly to
  prove conic / two-circle radial / pattern / gradient-anchored text /
  set_line_dash all honour the canvas2d spec (no CG degradations).

No behaviour change; identical TEST_CASE bodies + helpers (cg_make_bitmap,
CgPixelGrid).
